+++ dash_extensions/__init__.py
@@ -33,16 +33,26 @@ _this_module = _sys.modules[__name__]
 _js_dist = [
     {
         'relative_package_path': 'dash_extensions.min.js',
-
         'namespace': package_name
     },
     {
         'relative_package_path': 'dash_extensions.min.js.map',
-
         'namespace': package_name,
         'dynamic': True
     }
 ]
+for chunk in ["lottie", "mermaid", "burger"]:
+    _js_dist += [
+        {
+            'relative_package_path': f'async-{chunk}.js',
+            'namespace': package_name
+        },
+        {
+            'relative_package_path': f'async-{chunk}.js.map',
+            'namespace': package_name,
+            'dynamic': True
+        }
+    ]
 
 _css_dist = []