+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.59] - UNRELEASED
+
+### Added
+
+- Added `Purify` component to enable rendering of (sanitized) html.
+
+### Change
+
+- The `Mermaid` components now supports dynamic rendering (the component was rewritten completely from scratch).
+- Changed `Mermaid`, `Lottie`, and `Burger` components to use async loading. As a result, the size of the main `dash-extensions` bundle was reduced from > 1 MB to < 50 kB (!).
+
 ## [0.0.58] - 30-06-21
 
 ### Added